\name{GSwGBS-package}
\alias{GSwGBS-package}
\alias{GSwGBS}
\docType{package}
\title{
Genomic Selection with Genotyping-by-Sequencing
}
\description{
Contains functions for performing genomic selection with genotyping-by-sequencing markers obtained using the de novo method described by Poland et al. (2012). Included is a function for running the GBS SNP calling pipeline using Tassel 4.0. Functions for converting SNPs to numerically coded markers and performing multiple genomic selections methods using those markers are also included.
}
\details{
\tabular{ll}{
Package: \tab GSwGBS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-17\cr
License: \tab GPL-3\cr
}}
\author{
Chris Gaynor, Jesse Poland, Trevor Rife, and Allan Fritz
}
\references{
\cite{Bradbury, P.J., Z. Zhang, D.E. Kroon, T.M. Casstevens, Y. Ramdoss, and E.S. Buckler. 2007. TASSEL: Software for association mapping of complex traits in diverse samples. Bioinformatics 23:2633-2635.}

\cite{Poland, J.A., P.J. Brown, M.E. Sorrells, and J.-L. Jannink. 2012. Development of High-Density Genetic Maps for Barley and Wheat Using a Novel Two-Enzyme Genotyping-by-Sequencing Approach. PLoS ONE 7(2): e32253.}
}