\name{hap2marker}
\alias{hap2marker}
\title{
Hapfiles to Markers
}
\description{
Converts SNP data from the three ".hap" files outputted by \code{\link{GBSPipeline}} to a matrix of numerically coded markers for performing genomic selection.
}
\usage{
hap2marker(files = c("", "", ""), miss.max = 0.7, 
remove.blank = TRUE, n.core = "auto", 
impute.method = c("none", "median", "mean", "EM", "RF"), 
...)
}
\arguments{
  \item{files}{
A vector containing paths to three .hap files generated by \code{\link{GBSPipeline}}.
}
  \item{miss.max}{
Maximum allowable missing data points per SNP. This is an additional filter to the one performed by \code{\link{GBSPipeline}}, so it is possible to set stricter values without re-running the pipeline.
}
  \item{remove.blank}{
TRUE removes data labeled with the phrase "BLANK" in their name. This is intended to remove blank wells used to check plate orientation during DNA extraction.
}
  \item{n.core}{
Number of cores to use for parallel computing. If n.core="auto", n.core will be set to the number of detected cores. This may not be desirable due to memory requirements and/or overworking of CPU.
}
  \item{impute.method}{
Method to use for imputing missing markers. Only the method listed first is used.
}
  \item{\dots}{
Optional arguments passed to \code{\link[rrBLUP]{A.mat}} if using impute.method="EM" or "mean", or \code{\link[missForest]{missForest}} if using impute.method="RF".
}
}
\details{
Reads in data from three .hap files and converts SNPs to numeric coding. The dominate allele is changed to 1, the minor allele becomes -1, and heterzygotes become 0. After converting to numeric, markers with the same scoring are reduced to a single marker. For markers where each allele is equally frequent, a test is performed to determine if there are any markers with the same scoring after a sign change. Markers with the same scoring after a sign change are reduced to a single marker. Imputation of missing values is performed by the 'rrBLUP' package if method is "EM" or "mean" (see \code{\link[rrBLUP]{A.mat}}), the 'randomForest' package if method is "median" (see \code{\link[randomForest]{na.roughfix}}), or the 'missForest' package if method is "RF" (see \code{\link[missForest]{missForest}}).
}
\value{
Returns a list. The first item is a matrix where row names are line names taken from the first .hap file. Column names are M1, M2, ... The second item in the list is a data frame containing SNP sequence information. The first listed allele corresponds to markers coded as 1 and the second allele is coded as -1.
}
\examples{
\dontrun{
#Data to run this example is not included
tmp = hap2marker(files=c("hapFile1.hap", 
                         "hapFile2.hap", 
                         "hapFile3.hap"), 
                 impute.method="mean")
}
}